BlockPinner and ModPinner
These new tools allow you to select a block or mod, which will give you
a dedicated item for the selected block or mod.
The given items work very similar to the Pencil and Modifier, with the
key difference being that these items are dedicated to the block you
have selected.
Documentation will follow later.